pyproject.toml
[tool.poetry]
name = "wagtail-extended-search"
version = "0.1.0"
description = ""
authors = ["Cameron Lamb <jafacakes2011@hotmail.co.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
wagtail = "^6.3"
# OpenSearch
elasticsearch = "<7.14.0"         # highest version for OpenSearch support
elasticsearch-dsl = "^7.4.0"
django-elasticsearch-dsl = "^7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
